Add absolute links prefixed by a variable "langprefix"
Add a local Jekyll variable "langprefix" in each YAML front matter.
(Leave the variable empty for the canonical version. For the translated
versions, enter a path prefix in the style of "/en-US".

Copy phrases with absolute links.
On the copies, replace the language-dependent path prefixes by the
language-independent Liquid snippet "{{ page.langprefix }}".