Make function `extract_headline_id` more robust
In case the original MD files renders to different output
than expected ('<hN id="' with N in [0-9]), for example a
numberated heading like this

1. Awesome section
==================

a garbage id would be extracted without the check that
now is done by this commit. If the check fails then no
id will be returned by this function.